+++ service/src/main/kotlin/no/nav/su/se/bakover/service/søknad/lukk/LukkSøknadService.kt
@@ -3,7 +3,7 @@ package no.nav.su.se.bakover.service.søknad.lukk
 import arrow.core.Either
 import no.nav.su.se.bakover.domain.Sak
 import no.nav.su.se.bakover.domain.søknad.LukkSøknadRequest
-import no.nav.su.se.bakover.domain.søknadsbehandling.Søknadsbehandling
+import no.nav.su.se.bakover.domain.søknadsbehandling.KunneIkkeLukkeSøknadsbehandling
 
 interface LukkSøknadService {
     fun lukkSøknad(request: LukkSøknadRequest): Either<KunneIkkeLukkeSøknad, Sak>
@@ -16,7 +16,7 @@ sealed class KunneIkkeLukkeSøknad {
     object FantIkkePerson : KunneIkkeLukkeSøknad()
     object SøknadErAlleredeLukket : KunneIkkeLukkeSøknad()
     data class BehandlingErIFeilTilstand(
-        val feil: Søknadsbehandling.KunneIkkeLukkeSøknadsbehandling,
+        val feil: KunneIkkeLukkeSøknadsbehandling,
     ) : KunneIkkeLukkeSøknad()
 
     object UgyldigTrukketDato : KunneIkkeLukkeSøknad()
